Fix the lazy-list library module
There is one API braking change that is necessary to avoid possible
excessive memory consumption:  `number' is a constant that captures an
infinite lazy list of all the counting numbers; however, as a constant
it can never release the head of the list for garbage collection, so
memory used is proportional to the amount of the lazy list that is
forced/evaluated.  Thus, if one evaluates a large portion of the
infinity list, memory consumption goes toward infinity - essentially a
memory leak that lasts until the program is closed.  By making
`number()' a zero-arity function, proper use of the function can allow
for garbage collection of the head of the list as consumed.

In addition, I provide a `naturals()' function whis is just the above
function with the integer zero per-pended.

Essentially two major parts:

Part 1:  fix all functions so that there are no stack overflows for
infinity lazy lists when not necessary by the nature of the funciton.

Part 2:  Add the complete API of the non-lazy `List' library module so
that they behave in an equivalent way.